Refactor babel alias and imports
