+++ packages/lib-user/dev/index.js
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client'
 
-import App from './components/App/index.js'
+import App from './components/App'
 
 function getQueryParams() {
   if (window.location) {
