+++ packages/lib-user/package.json
@@ -26,7 +26,7 @@
     "dev": "BABEL_ENV=es6 webpack serve --config webpack.dev.js",
     "eslint": "eslint ./src --fix | snazzy",
     "prepare": "yarn build",
-    "test": "mocha --config ./test/.mocharc.json ./.storybook/specConfig.js \"./src/**/*.spec.js\"",
+    "test": "NODE_ENV=test mocha --config ./test/.mocharc.json ./.storybook/specConfig.js \"./src/**/*.spec.js\"",
     "test:ci": "mocha --config ./test/.mocharc.json ./.storybook/specConfig.js --reporter=min \"./src/**/*.spec.js\"",
     "storybook": "storybook dev -p 6008",
     "build-storybook": "storybook build"
@@ -60,7 +60,6 @@
     "@testing-library/react": "~14.2.0",
     "@testing-library/user-event": "~14.5.0",
     "babel-loader": "~9.1.0",
-    "babel-plugin-module-resolver": "~5.0.0",
     "babel-plugin-transform-imports": "~2.0.0",
     "chai": "~4.3.4",
     "chai-dom": "~1.12.0",
