+++ packages/lib-user/src/components/shared/ContentBox/index.js
@@ -1,1 +1,1 @@
-export { default as ContentBox } from './ContentBox.js'
+export { default } from './ContentBox.js'
