+++ packages/lib-user/src/components/shared/Layout/index.js
@@ -1,1 +1,1 @@
-export { default as Layout } from './Layout.js'
+export { default } from './Layout.js'
