+++ packages/lib-user/src/components/shared/ProfileHeader/index.js
@@ -1,1 +1,1 @@
-export { default as ProfileHeader } from './ProfileHeader.js'
+export { default } from './ProfileHeader.js'
