+++ packages/lib-user/src/components/shared/ProjectCard/index.js
@@ -1,1 +1,1 @@
-export { default as ProjectCard } from './ProjectCard.js'
+export { default } from './ProjectCard.js'
