+++ packages/lib-user/src/components/shared/Select/index.js
@@ -1,1 +1,1 @@
-export { default as Select } from './Select.js'
+export { default } from './Select.js'
