+++ packages/lib-user/src/components/shared/Tabs/index.js
@@ -1,1 +1,1 @@
-export { default as Tabs } from './Tabs.js'
+export { default } from './Tabs.js'
