+++ packages/lib-user/src/components/shared/index.js
@@ -1,7 +1,7 @@
-export { default as BarChart } from './BarChart/BarChart.js'
-export { default as ContentBox } from './ContentBox/ContentBox.js'
-export { default as Layout } from './Layout/Layout.js'
-export { default as ProfileHeader } from './ProfileHeader/ProfileHeader.js'
-export { default as ProjectCard } from './ProjectCard/ProjectCard.js'
-export { default as Select } from './Select/Select.js'
-export { default as Tabs } from './Tabs/Tabs.js'
+export { default as BarChart } from './BarChart'
+export { default as ContentBox } from './ContentBox'
+export { default as Layout } from './Layout'
+export { default as ProfileHeader } from './ProfileHeader'
+export { default as ProjectCard } from './ProjectCard'
+export { default as Select } from './Select'
+export { default as Tabs } from './Tabs'
