+++ packages/lib-user/src/hooks/index.js
@@ -1,7 +1,7 @@
-export { default as useGroupStats } from './useGroupStats.js'
-export { default as usePanoptesAuth } from './usePanoptesAuth.js'
-export { default as usePanoptesMemberships } from './usePanoptesMemberships.js'
-export { default as usePanoptesProjects } from './usePanoptesProjects.js'
-export { default as usePanoptesUser } from './usePanoptesUser.js'
-export { default as usePanoptesUserGroup } from './usePanoptesUserGroup.js'
-export { default as useUserStats } from './useUserStats.js'
+export { useGroupStats } from './useGroupStats.js'
+export { usePanoptesAuth } from './usePanoptesAuth.js'
+export { usePanoptesMemberships } from './usePanoptesMemberships.js'
+export { usePanoptesProjects } from './usePanoptesProjects.js'
+export { usePanoptesUser } from './usePanoptesUser.js'
+export { usePanoptesUserGroup } from './usePanoptesUserGroup.js'
+export { useUserStats } from './useUserStats.js'
