+++ packages/lib-user/src/index.js
@@ -1,30 +1,30 @@
 // components
-export { default as UserStats } from './components/UserStats/index.js'
-export { default as MyGroups } from './components/MyGroups/index.js'
-export { default as GroupStats } from './components/GroupStats/index.js'
+export { default as UserStats } from './components/UserStats'
+export { default as MyGroups } from './components/MyGroups'
+export { default as GroupStats } from './components/GroupStats'
 
 // components/shared
-export { default as BarChart } from './components/shared/BarChart/index.js'
-export { default as ContentBox } from './components/shared/ContentBox/index.js'
-export { default as Layout } from './components/shared/Layout/index.js'
-export { default as ProfileHeader } from './components/shared/ProfileHeader/index.js'
-export { default as ProjectCard } from './components/shared/ProjectCard/index.js'
-export { default as Select } from './components/shared/Select/index.js'
-export { default as Tabs } from './components/shared/Tabs/index.js'
+export { default as BarChart } from './components/shared/BarChart'
+export { default as ContentBox } from './components/shared/ContentBox'
+export { default as Layout } from './components/shared/Layout'
+export { default as ProfileHeader } from './components/shared/ProfileHeader'
+export { default as ProjectCard } from './components/shared/ProjectCard'
+export { default as Select } from './components/shared/Select'
+export { default as Tabs } from './components/shared/Tabs'
 
 // hooks
-export { default as useGroupStats } from './hooks/useGroupStats.js'
-export { default as usePanoptesAuth } from './hooks/usePanoptesAuth.js'
-export { default as usePanoptesMemberships } from './hooks/usePanoptesMemberships.js'
-export { default as usePanoptesProjects } from './hooks/usePanoptesProjects.js'
-export { default as usePanoptesUser } from './hooks/usePanoptesUser.js'
-export { default as usePanoptesUserGroup } from './hooks/usePanoptesUserGroup.js'
-export { default as useUserStats } from './hooks/useUserStats.js'
+export { useGroupStats } from './hooks/useGroupStats.js'
+export { usePanoptesAuth } from './hooks/usePanoptesAuth.js'
+export { usePanoptesMemberships } from './hooks/usePanoptesMemberships.js'
+export { usePanoptesProjects } from './hooks/usePanoptesProjects.js'
+export { usePanoptesUser } from './hooks/usePanoptesUser.js'
+export { usePanoptesUserGroup } from './hooks/usePanoptesUserGroup.js'
+export { useUserStats } from './hooks/useUserStats.js'
 
 // utils
-export { default as createPanoptesUserGroup } from './utils/createPanoptesUserGroup.js'
-export { default as dateRanges } from './utils/dateRanges.js'
-export { default as deletePanoptesUserGroup } from './utils/deletePanoptesUserGroup.js'
-export { default as getBearerToken } from './utils/getBearerToken.js'
-export { default as getDateInterval } from './utils/getDateInterval.js'
-export { default as updatePanoptesUserGroup } from './utils/updatePanoptesUserGroup.js'
+export { createPanoptesUserGroup } from './utils/createPanoptesUserGroup.js'
+export { dateRanges } from './utils/dateRanges.js'
+export { deletePanoptesUserGroup } from './utils/deletePanoptesUserGroup.js'
+export { getBearerToken } from './utils/getBearerToken.js'
+export { getDateInterval } from './utils/getDateInterval.js'
+export { updatePanoptesUserGroup } from './utils/updatePanoptesUserGroup.js'
