+++ packages/lib-user/src/utils/index.js
@@ -1,6 +1,6 @@
-export { default as createPanoptesUserGroup } from './createPanoptesUserGroup.js'
-export { default as dateRanges } from './dateRanges.js'
-export { default as deletePanoptesUserGroup } from './deletePanoptesUserGroup.js'
-export { default as getBearerToken } from './getBearerToken.js'
-export { default as getDateInterval } from './getDateInterval.js'
-export { default as updatePanoptesUserGroup } from './updatePanoptesUserGroup.js'
+export { createPanoptesUserGroup } from './createPanoptesUserGroup.js'
+export { dateRanges } from './dateRanges.js'
+export { deletePanoptesUserGroup } from './deletePanoptesUserGroup.js'
+export { getBearerToken } from './getBearerToken.js'
+export { getDateInterval } from './getDateInterval.js'
+export { updatePanoptesUserGroup } from './updatePanoptesUserGroup.js'
