+++ packages/lib-user/webpack.dev.js
@@ -40,11 +40,6 @@ module.exports = {
   ],
   mode: 'development',
   resolve: {
-    alias: {
-      '@components': path.resolve(__dirname, 'src/components'),
-      '@hooks': path.resolve(__dirname, 'src/hooks'),
-      '@utils': path.resolve(__dirname, 'src/utils')
-    },
     fallback: {
       fs: false,
       // for markdown-it plugins
